+++ src/utils/types/types.ts
@@ -1,0 +1,97 @@
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export enum CalendarLength {
+  NULL = 0,
+  LESS_THAN_ONE_DAY = 1,
+  ONE_TO_TWO_DAYS = 2,
+  THREE_TO_SEVEN_DAYS = 3,
+  TWO_TO_THREE_WEEKS = 4,
+  FOUR_TO_ELEVEN_WEEKS = 5,
+  THREE_TO_FIVE_MONTHS = 6,
+  SIX_TO_TWELVE_MONTHS = 7,
+  THIRTEEN_MONTHS_TO_TWO_YEARS = 8,
+  THREE_TO_FOUR_YEARS = 9,
+  MORE_THAN_FOUR_YEARS = 10,
+}
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6;
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type ColumnSize = 1 | 2 | 3 | 4 | 5 | 6;
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type ThemeColors = "navy" | "blue" | "green" | "purple" | "orange";
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type Elements =
+  | "div"
+  | "section"
+  | "article"
+  | "aside"
+  | "header"
+  | "footer"
+  | "main"
+  | "nav"
+  | "span"
+  | "p"
+  | "h1"
+  | "h2"
+  | "h3"
+  | "h4"
+  | "h5"
+  | "h6"
+  | "ul"
+  | "ol"
+  | "li";
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type IconWeight =
+  | "regular"
+  | "bold"
+  | "thin"
+  | "light"
+  | "duotone"
+  | "fill";
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type SectionIcons =
+  | "explore"
+  | "jobs"
+  | "support"
+  | "training"
+  | "exploreBold"
+  | "jobsBold"
+  | "supportBold"
+  | "trainingBold";
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type BreakNames =
+  | "none"
+  | "xxs"
+  | "xs"
+  | "sm"
+  | "md"
+  | "lg"
+  | "xl"
+  | "xxl";
+
+//////////////////////////////////////////////////////
+//////////////////////////////////////////////////////
+
+export type Gaps = BreakNames | "none" | "micro" | "mega";
